+++ server/integrations/twitter.js
@@ -18,23 +18,27 @@ log.debug(`Following: ${follow.join(',')}`)
 var client = new Twitter(config.get('twitter.auth'))
 
 function transformTweet (enrichedTweet) {
-  return JSON.stringify({
-    text: twitterText.autoLink(twitterText.htmlEscape(getTweetText(enrichedTweet.tweet))),
+  return {
+    text: twitterText.autoLink(
+      twitterText.htmlEscape(getTweetText(enrichedTweet.tweet))
+    ),
     created_at: enrichedTweet.tweet.created_at,
     displayName: enrichedTweet.tweet.user.name,
     username: enrichedTweet.tweet.user.screen_name,
     userProfilePic: enrichedTweet.tweet.user.profile_image_url_https,
     sentiment: enrichedTweet.sentiment,
     retweeted: enrichedTweet.tweet.retweeted
-  })
+  }
 }
 
 function getTweetText (tweet) {
-  return tweet.truncated && tweet.extended_tweet ? tweet.extended_tweet.full_text : tweet.text
+  return tweet.truncated && tweet.extended_tweet
+    ? tweet.extended_tweet.full_text
+    : tweet.text
 }
 
 function getTweetSentiment (tweet) {
-  return awsComprehend.detectSentiment(tweet.text).then((sentiment) => {
+  return awsComprehend.detectSentiment(tweet.text).then(sentiment => {
     return {
       id: tweet.id_str,
       created_at: new Date(tweet.created_at),
@@ -45,9 +49,12 @@ function getTweetSentiment (tweet) {
 }
 
 function persistTweet (enrichedTweet) {
-  return mongodb.connect.then((db) =>
-    db.collection('tweets')
-      .insertOne(enrichedTweet).then(() => enrichedTweet))
+  return mongodb.connect.then(db =>
+    db
+      .collection('tweets')
+      .insertOne(enrichedTweet)
+      .then(() => enrichedTweet)
+  )
 }
 
 function persistTweetUserMentions (enrichedTweet) {
@@ -66,24 +73,34 @@ function handleTweet (tweet) {
     .then(persistTweet)
     .then(persistTweetUserMentions)
     .then(transformTweet)
-    .then((transformedTweet) => {
-      listeners.forEach((listener) => listener(transformedTweet))
+    .then(transformedTweet => {
+      transformedTweet = JSON.stringify(transformedTweet)
+      listeners.forEach(listener => listener(transformedTweet))
       return transformedTweet
-    }).catch((err) => {
+    })
+    .catch(err => {
       log.error(`Could not process tweet [${tweet.id}]`, err)
     })
   return lastTweet
 }
 
 function getLatestTweet () {
-  lastTweet = client.get('search/tweets', { q: `${topics.concat(follow).map((item) => encodeURIComponent(item)).join(' OR ')}`, result_type: 'recent', count: 1 })
-    .then((tweets) => {
+  lastTweet = client
+    .get('search/tweets', {
+      q: `${topics
+        .concat(follow)
+        .map(item => encodeURIComponent(item))
+        .join(' OR ')}`,
+      result_type: 'recent',
+      count: 1
+    })
+    .then(tweets => {
       if (!tweets.statuses || tweets.statuses.length === 0) {
         throw new Error('Did not retrieve any tweets. Will try again.')
       }
       return handleTweet(tweets.statuses[0])
     })
-    .catch((err) => {
+    .catch(err => {
       log.error('Error during getLatestTweet', err)
       setTimeout(getLatestTweet, 5000)
     })
@@ -96,18 +113,20 @@ const streamBackoff = backoff.exponential({
 
 function startStream () {
   log.info('Connection to Twitter stream')
-  var stream = client.stream('statuses/filter', { track: topics.concat(follow).join(',') })
+  var stream = client.stream('statuses/filter', {
+    track: topics.concat(follow).join(',')
+  })
 
   stream.on('response', () => {
     log.info('Connected to Twitter stream')
     streamBackoff.reset()
   })
 
-  stream.on('data', (event) => {
+  stream.on('data', event => {
     handleTweet(event)
   })
 
-  stream.on('error', (err) => {
+  stream.on('error', err => {
     log.warn('Error connecting to Twitter stream', err)
   })
 
@@ -125,8 +144,10 @@ streamBackoff.on('ready', (number, delay) => {
 getLatestTweet()
 startStream()
 
-module.exports.addListener = (listener) => {
+module.exports.addListener = listener => {
   listeners.push(listener)
 }
 
 module.exports.getLatestTweet = () => lastTweet
+
+module.exports.transformTweet = transformTweet